Update 2023-05-16-a-teacher.md
Based on KA chatgpt4 with prompts of make the following blog entry a little bit more exciting. The audience is students, parents and sometimes teachers. Given I'm a new teacher, it needs to be written with a bit of humility

Hm. I don't like talking about myself that much. While I am excited, it is more that by being present and intentional with them, I can impart some character development and perhaps lead many to learn computer science enough to feel in-the-zone when solving an interesting problem.